UIBULKED-115: In-app bulk edit form - holdings location
